+++ src/main/java/com/example/solid_classes/core/product_variation/dto/ProductVariationResponseDto.java
@@ -1,5 +1,6 @@
 package com.example.solid_classes.core.product_variation.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import lombok.Builder;
@@ -8,7 +9,7 @@ import lombok.Builder;
 public class ProductVariationResponseDto {
     private UUID id;
     private String variationValue;
-    private Double variationPrice;
+    private BigDecimal variationPrice;
     private String variationCategoryName;
     private String variationProductName;
 }
